Consolidates content compiler tests into one test suite class.
