+++ Test/ContentPipeline/ContentCompilerTest.cs
@@ -17,7 +17,7 @@ using System.Reflection;
 
 namespace MonoGame.Tests.ContentPipeline
 {
-    class ArrayContentCompilerTest
+    class ContentCompilerTest
     {
         class TestDataClass : IEquatable<TestDataClass>
         {
@@ -221,6 +221,48 @@ namespace MonoGame.Tests.ContentPipeline
                         for (int x = 0; x < expected.GetLength(0); x++)
                             Assert.AreEqual(expected[x, y, z], result[x, y, z]);
             });
+        }
+
+        [Test]
+        public void ShouldSerializePropertyGivenNameOfItem()
+        {
+            var expected = new HasNoIndexer { Item = "lorem-ipsum" };
+
+            TestCompiler.CompileAndLoadAssets(expected, result =>
+            {
+                Assert.AreEqual(expected.Item, result.Item);
+            });
+        }
+
+        class HasNoIndexer
+        {
+            public string Item { get; set; }
+        }
+
+        [Test]
+        public void ShouldNotSerializePropertyGivenIndexer()
+        {
+            var expected = new HasIndexer();
+            expected["anything"] = "value";
+
+            TestCompiler.CompileAndLoadAssets(expected, result =>
+            {
+                Assert.AreNotEqual(expected["anything"], result["anything"]);
+            });
+        }
+
+        class HasIndexer
+        {
+            readonly Dictionary<string, string> _dictionary = new Dictionary<string, string>();
+            public string this[string key]
+            {
+                get
+                { 
+                    string value;
+                    return _dictionary.TryGetValue(key, out value) ? value : null;
+                }
+                set { _dictionary[key] = value; }
+            }
         }
     }
 }
