--- Test/ContentPipeline/XmlContentCompilerTest.cs
@@ -1,50 +1,0 @@
-﻿using NUnit.Framework;
-using System.Collections.Generic;
-
-namespace MonoGame.Tests.ContentPipeline
-{
-    class XmlContentCompilerTest
-    {
-        [Test]
-        public void ShouldSerializePropertyGivenNameOfItem()
-        {
-            var expected = new HasNoIndexer { Item = "lorem-ipsum" };
-
-            TestCompiler.CompileAndLoadAssets(expected, result =>
-            {
-                Assert.AreEqual(expected.Item, result.Item);
-            });
-        }
-
-        class HasNoIndexer
-        {
-            public string Item { get; set; }
-        }
-
-        [Test]
-        public void ShouldNotSerializePropertyGivenIndexer()
-        {
-            var expected = new HasIndexer();
-            expected["anything"] = "value";
-
-            TestCompiler.CompileAndLoadAssets(expected, result =>
-            {
-                Assert.AreNotEqual(expected["anything"], result["anything"]);
-            });
-        }
-
-        class HasIndexer
-        {
-            readonly Dictionary<string, string> _dictionary = new Dictionary<string, string>();
-            public string this[string key]
-            {
-                get
-                { 
-                    string value;
-                    return _dictionary.TryGetValue(key, out value) ? value : null;
-                }
-                set { _dictionary[key] = value; }
-            }
-        }
-    }
-}
